requirements_cgan.txt
# Conditional GAN Training Requirements

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data & Utilities
numpy>=1.24.0
pandas>=1.5.0
pillow>=9.5.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: Model evaluation
# scikit-learn>=1.2.0  # For FID/IS metrics
# scipy>=1.10.0       # For FID computation
